[feat] 단일 구독은 완료

+++ src/main/java/com/example/monthlylifebackend/subscribe/repository/SubscribeRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SubscribeRepository extends JpaRepository<Subscribe, Integer> {
+public interface SubscribeRepository extends JpaRepository<Subscribe, Long> {
     @Query("""
     SELECT new com.example.monthlylifebackend.subscribe.dto.response.GetDeliveryListRes(
         s.idx,
